feat: add static files serving

+++ backend/src/users/users.controller.ts
@@ -9,6 +9,8 @@ import {
 	ParseIntPipe,
 	UseInterceptors,
 	UploadedFile,
+	ForbiddenException,
+	BadRequestException,
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { UsersService } from './users.service';
@@ -16,6 +18,9 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { diskStorage } from 'multer';
+import { Role, User } from '@prisma/client';
+import { GetUser } from 'src/decorators/get-user.decorator';
+import { Roles } from 'src/decorators/role.decorator';
 import { generateFilename, imageFileFilter } from 'src/helpers/files-helper';
 
 @ApiTags('Users')
@@ -33,6 +38,7 @@ export class UsersController {
 		return this.userService.findAll();
 	}
 
+	@Roles([Role.ADMIN])
 	@Post()
 	createUser(@Body() createUserDto: CreateUserDto) {
 		return this.userService.create(createUserDto);
@@ -42,7 +48,7 @@ export class UsersController {
 	@UseInterceptors(
 		FileInterceptor('image', {
 			storage: diskStorage({
-				destination: './public/avatars',
+				destination: './uploads/avatar',
 				filename: generateFilename,
 			}),
 			fileFilter: imageFileFilter,
@@ -50,28 +56,39 @@ export class UsersController {
 	)
 	setUserAvatar(
 		@Param('id', ParseIntPipe) id: number,
-		@UploadedFile() // new ParseFilePipe({
-		image // 	validators: [
-		// 		new MaxFileSizeValidator({ maxSize: 1048576 }),
-		// 		new FileTypeValidator({ fileType: '(png|jpeg|jpg)' }),
-		// 	],
-		// 	errorHttpStatusCode: HttpStatus.UNPROCESSABLE_ENTITY,
-		// }),
-		: Express.Multer.File,
+		@UploadedFile() file: Express.Multer.File,
+		@GetUser() user: User,
 	) {
-		return this.userService.setAvatar(id, image.path);
+		if (user.id !== id && user.role !== Role.ADMIN) {
+			throw new ForbiddenException('Cannot set user avatar');
+		}
+
+		if (!file) {
+			throw new BadRequestException('Invalid file');
+		}
+
+		console.log(file.path);
+
+		return this.userService.setAvatar(id, file.originalname);
 	}
 
 	@Patch(':id')
 	updateUser(
 		@Param('id', ParseIntPipe) id: number,
 		@Body() updateUserDto: UpdateUserDto,
+		@GetUser() user: User,
 	) {
+		if (user.id !== id && user.role !== Role.ADMIN) {
+			throw new ForbiddenException('Cannot update user');
+		}
 		return this.userService.update(id, updateUserDto);
 	}
 
 	@Delete(':id')
-	deleteUser(@Param('id', ParseIntPipe) id: number) {
+	deleteUser(@Param('id', ParseIntPipe) id: number, @GetUser() user: User) {
+		if (user.id !== id && user.role !== Role.ADMIN) {
+			throw new ForbiddenException('Cannot delete user');
+		}
 		return this.userService.delete(id);
 	}
 }
